Only change user format preference on request
If an update_user call didn't specify a preference for email format
(T or H), Basket was defaulting to 'H', then setting that in ET
even if the user had previously selected 'T'.

Instead, don't send ET a format if the API call didn't specify one.
We still use the user's previous preference, if any, for sending
their welcome messages, or else HTML.