Bug 865058: Add order field to newsletters
Add an `order` field to newsletters. Bedrock already has code on the
newsletter preference center page so that if the newsletters include
an `order` field, they'll be sorted by that for display, instead of
by the title.

Also update the admin config to make it easier to work with the order
fields, by making them editable right in the display list.  Add a
few other fields to that for good measure.