Remove timing of retried tasks
Due to the fact that the retries are delayed by design
the timing numbers for these retried tasks are distorting
our view of how the queue is actually functioning. Because
the first part of the et_task function pops "start_time" from
kwargs the retries will no longer be timed at all.