Flip SFDC test data switch for testing email domains
There is a flag on our SFDC records that indicates that
it was added as a User Acceptance Test. We can always flip
this flag for known testing email address domains. By default
this will be restmail.net (mozilla's own email testing service)
and example.com which won't actually do anything within SFDC
or SFMC regardless.