Fix email missing in common-voice task retries
The task was modifying the data dict passed to it (removed the email
key). This meant that the data passed to the task on a retry due to a
failure of the SFDC call would be missing the email and would fail with
a KeyError. This change fixes that and adds tests for the task.