Add CTMSInterface class
CTMSInterface has methods for each of the public CTMS API calls.

It is difficult to test at this level without a lot of mocking or
simulating CTMS, and I tried a few approaches that made the testing code
as complex as the application code.

mock_interface() is a compromise, which will allow mocking of successful
responses as well as errors. Integrated testing will be needed to fully
test the code.