Drop pip-tools from the Dockerfile and rebuild requirements on the host, not in the container
Retains Makefile commands as the primary invocation route - make compile-requirements and make upgrade-requirements

Also note that the makefile installs the latest pip-tools ahead of use, rather than a pinned and poss unpatched version. (Dependency bots likely won't spot a pinning in a Makefile)

This change was needed because in-container compilation was failing and various pip+pip-tools versions attempted clashed due to
```
TypeError: make_requirement_preparer() got an unexpected keyword argument 'wheel_download_dir'
```
which may have been a chicken-and-egg problem, but was not worth solving when a simpler route was around.