Refactor to run pip-compile-multi within the docker container
This gives us guaranteed install stability regardless of the source platform being used to run the update

Note that pip-tools and pip-compile-multi are installed ONLY for compiling or upgrading requirements, rather than adding them to the main build.
pip-tools needed pinning to a recent version to get around incompatibility issues with pip 21 and pip-tools.